Move cpp files and use absolute imports
