+++ main.cpp
@@ -39,40 +39,41 @@
 #include <QScreen>
 #include <QRegExp>
 #include <QThread>
-#include "clipboardAdapter.h"
-#include "filter.h"
-#include "oscursor.h"
-#include "oshelper.h"
-#include "WalletManager.h"
-#include "Wallet.h"
-#include "QRCodeImageProvider.h"
-#include "PendingTransaction.h"
-#include "UnsignedTransaction.h"
-#include "TranslationManager.h"
-#include "TransactionInfo.h"
-#include "TransactionHistory.h"
-#include "model/TransactionHistoryModel.h"
-#include "model/TransactionHistorySortFilterModel.h"
-#include "AddressBook.h"
-#include "model/AddressBookModel.h"
-#include "Subaddress.h"
-#include "model/SubaddressModel.h"
-#include "SubaddressAccount.h"
-#include "model/SubaddressAccountModel.h"
+
 #include "wallet/api/wallet2_api.h"
-#include "Logger.h"
-#include "MainApp.h"
-#include "qt/ipc.h"
-#include "qt/utils.h"
-#include "qt/mime.h"
+#include "src/libwalletqt/WalletManager.h"
+#include "src/libwalletqt/Wallet.h"
+#include "src/libwalletqt/QRCodeImageProvider.h"
+#include "src/libwalletqt/PendingTransaction.h"
+#include "src/libwalletqt/UnsignedTransaction.h"
+#include "src/libwalletqt/TransactionInfo.h"
+#include "src/libwalletqt/TransactionHistory.h"
+#include "src/libwalletqt/AddressBook.h"
+#include "src/libwalletqt/Subaddress.h"
+#include "src/libwalletqt/SubaddressAccount.h"
+#include "src/model/AddressBookModel.h"
+#include "src/model/SubaddressModel.h"
+#include "src/model/SubaddressAccountModel.h"
+#include "src/model/TransactionHistoryModel.h"
+#include "src/model/TransactionHistorySortFilterModel.h"
+#include "src/qt/clipboardAdapter.h"
+#include "src/qt/filter.h"
+#include "src/qt/oscursor.h"
+#include "src/qt/oshelper.h"
+#include "src/qt/TranslationManager.h"
+#include "src/qt/Logger.h"
+#include "src/qt/MainApp.h"
+#include "src/qt/ipc.h"
+#include "src/qt/utils.h"
+#include "src/qt/mime.h"
 #include "src/qt/TailsOS.h"
 #include "src/qt/KeysFiles.h"
 #include "src/qt/MoneroSettings.h"
-#include "qt/prices.h"
+#include "src/qt/prices.h"
 
 // IOS exclusions
 #ifndef Q_OS_IOS
-#include "daemon/DaemonManager.h"
+#include "src/daemon/DaemonManager.h"
 #endif
 
 #ifdef WITH_SCANNER
