+++ src/qt/Logger.h
