+++ src/qt/MainApp.h
