+++ src/qt/TranslationManager.h
