+++ src/qt/clipboardAdapter.h
