+++ src/qt/filter.cpp
@@ -26,10 +26,11 @@
 // STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF
 // THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-#include "filter.h"
 #include <QtGlobal>
 #include <QKeyEvent>
 #include <QDebug>
+
+#include "src/qt/filter.h"
 
 #ifdef QT_DEBUG
     #include "private/qabstractanimation_p.h"
