+++ src/qt/filter.h
