+++ src/qt/oscursor.h
