+++ src/qt/oshelper.h
